Adapted evaluator "Exploration" for Dungeon Crawler (Only Interior)
Created:
Class DCExploration
Interface IContextualEvaluator

DCExploration is an evaluator that equalizes distances between player and "points of interest" as chests, weapons, food, etc. It calculates distances with a Flood Fill algorithm that considers the first visible Interior Layer and check for doors before traversing to a different zone. If there is not a visible Interior Layer, distances are calculated with Manhattan method.

Future Work:
- Flood Fill algorithm should consider interior walls of a zone and elements with collider characteristic.
- The evaluator should consider layers indicated in the contextual menu of the Population Assistant Window.
- The evaluator should be able to consider and interpret an Exterior Layer.
- The evaluator should be able to consider more than one layer at a time.