RectExtensions Class + AutoSelect Area Bugfix
Added a class named "RectExtensions", just so i could add a function called "GetCombinedArea". That way i made the code cleaner within PopulationAssistantWindow. Also fixed a bug with the auto select button, which considered x=0 and y=0 in the calculations of area, causing a selection of an unwanted zone when the elements were painted in negative positions.