Multi-layer connections painting option for Interiors
If the option is enabled, when painting a door or window connection in a selected interior layer, the same connection will be painted in another interior layer at the same position if a previous non-empty connection exists. This provides a faster way to, for example, make connections between zones from different layers.