complete fix in quest assistant
there were issues with getting the GUID from bundles as it is now allowed on threads so instead GUIDS are stored directly and used for hashing compare alongside the tile coordinates. Replaced all Unity.Random for System.Random in threaded functions. Fixed an issue in the toolbar that would disable it instead of disabling the thread process.